Move configuration handling to CodeGenerator
